requirements.txt
# Web framework
django==4.2.16
djangorestframework==3.15.2
gunicorn==22.0.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Cache / Celery broker+backend
redis==5.1.1
django-redis==5.4.0
celery[redis]==5.3.6

# AI summarization (Claude, ChatGPT, or Gemini)
anthropic==0.40.0
openai==1.54.4
google-generativeai==0.8.3

# HTTP client (job fetcher)
httpx==0.27.2

# Config
python-dotenv==1.0.1
